+++ src/main/java/com/izbicki/jakub/Repository/UserRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -21,5 +22,5 @@ public interface UserRepository extends CrudRepository<User, Long> {
     @Modifying
     @Transactional
     @Query("UPDATE User SET rentalWallet = :walletValue WHERE login = :login")
-    void updateWallet(@Param("login") String login, @Param("walletValue") float walletValue);
+    void updateWallet(@Param("login") String login, @Param("walletValue") BigDecimal walletValue);
 }
